Use ELF Section Headers to Compute ELF Size
Previously, we were using program headers to compute total size of ELF
memory when loaded in memory. Unfortunately, this is not enough, as some
segments that are not loaded may also allocate memory, such as the BSS.

This commit fixes this problem by changing this logic to use section
headers instead.

Signed-off-by: Pedro Henrique Penna <ppenna@microsoft.com>