[host/{mem/layout,sandbox/mem_mgr}] removes statically defined mem layout + removes MemMgrWrapper
This commit completely removes the file w/ the mem layout. There are still
references to it around, but those will be removed in subsequent commits.

With the upcoming refactors, there is no need for the MemMgrWrapper abstraction. This
commit removes in in favour of directly utilizing a SandboxMemoryManager<S> in subsequent
commits.

Signed-off-by: danbugs <danilochiarlone@gmail.com>