[hyperlight_component_util] Support passing host resources to guest functions
Previously, the guest could acquire and use host resources through
host function calls, but the original call from the host into the
guest could not pass in owned or borrowed host resources.  This commit
fixes in a number of bugs in order to make that possible, and adds
rudimentary tests to witguest.

Signed-off-by: Lucy Menon <168595099+syntactically@users.noreply.github.com>