[nix] Apply binutils wrapper to Rust's ld.lld
In Rust 1.90, rustc defaults to -fuse-ld=lld, which in turn uses a
bundled-with-the-compiler ld.lld driver. That driver was not being
properly wrapped, leading to generated executables failing to find
libraries due to missing DT_RUNPATH entries. This commit modifies the
mozilla Rust overlay to wrap the ld.lld driver.

Signed-off-by: Simon Davies <simongdavies@users.noreply.github.com>