Remove the last vestiges of autodetecting heap/stack size from exes
When we supported loading guest binaries from PE files, we also
supported autodetecting reasonable defaults for heap and stack size
for the sandbox based on the loaded executable's stack/heap size
hints.  ELF files did not have a similar convention for stack/heap
size hints embedded in the file, so when loading an ELF file, we just
used some vaguely reasonable small defaults (64k stack and 128k heap).
Now that PE support is gone, these are in fact the defaults used for
/all/ files, so it doesn't make much sense that we bother putting them
into an executable information structure and passing it around.  This
commit removes that vestigial use of the ExeInfo structure, replacing
it with defaults inline in `sandbox::config`.

Signed-off-by: Simon Davies <simongdavies@users.noreply.github.com>