Which came first, the snapshot or the sandbox?
This commit changes the Hyperlight API so that every sandbox is
created from a snapshot. This is useful for several reasons; most
immediately, in the same commit, note that it allows us to avoid
precommitting to a size for the page table region, so we no longer
need to estimate that region's size.