Ensure that processor control structures stay accessible after snapshots
Previously, the amd64 TSS was written directly into a page of scratch
memory, which would not be noticed by the snapshotting process and
would result in the TSS getting lost. This commit changes the code
that sets up all these processor control structures to map the
allocated pages to a new address first, in order to ensure that they
are included in the snapshot.  It also refactors the
architecture-specific initialisation code somewhat, in order to make
it a bit easier to have one memory allocation shared by all the
relevant tables.