Unify page table manipulation code between the guest and the host (#1093)
* Unify page table manipulation code between the guest and the host

Currently, the guest and the host both have code that manipulates
architecture-specific page table structures: the guest has a general
map operation, and the host has a much more specific routine that
builds an identity map.  As we move to more complex virtual memory
configurations in the guest, the host will need the ability to build
more complex mappings in the guest, so this commit removes the simple
implementation in the host, and replaces it with calls to the
implementation originally written for the guest (now moved to
`hyperlight_common` and factored into an architecture-independent
interface and architecture-specific code parts).

Signed-off-by: Simon Davies <simongdavies@users.noreply.github.com>

* Only include vm mod when init-paging is enabled

Signed-off-by: Simon Davies <simongdavies@users.noreply.github.com>

* Review changes

Signed-off-by: Simon Davies <simongdavies@users.noreply.github.com>

* last ever fix

Signed-off-by: Simon Davies <simongdavies@users.noreply.github.com>

---------

Signed-off-by: Simon Davies <simongdavies@users.noreply.github.com>
Co-authored-by: Lucy Menon <168595099+syntactically@users.noreply.github.com>